feat(plugin): filter for discord & vesktop & web
This filtering shows all the valid plugins for the chosen platform,
instead of plugins which only work on the chosen platform like it did
before.

This is under the assumption that plugins with the target flag discord
require the official discord client, those with the desktop flag only on
vesktop and those with web only on the webclient.